refactor: piece 패키지 생성으로 관련 클래스 패키지 구조 수정
