--- src/main/java/chess/domain/Piece.java
@@ -1,5 +1,0 @@
-package chess.domain;
-
-public interface Piece {
-    boolean move(Position newPosition, Board board);
-}
