+++ src/main/java/chess/domain/piece/ChessPiece.java
@@ -1,5 +1,8 @@
-package chess.domain;
+package chess.domain.piece;
 
+import chess.domain.Board;
+import chess.domain.PieceInfo;
+import chess.domain.Position;
 import chess.domain.strategy.MoveStrategy;
 
 public abstract class ChessPiece implements Piece {
