+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+import chess.domain.Board;
+import chess.domain.Position;
+
+public interface Piece {
+    boolean move(Position newPosition, Board board);
+}
